+++ index.html
@@ -10,7 +10,7 @@
 
   <script type="text/javascript">
     SearchWidget.new({
-      endpoint: "http://rrsoft.apib.dev/search/sample_api"
+      endpoint: ""
     })
   </script>
 </html>
