+++ public/index.html
@@ -1,0 +1,16 @@
+<html>
+  <head>
+    <title>zApiness</title>
+  </head>
+  <body>
+    <div class="app" id="root">
+    </div>
+  </body>
+  <script src="/static/bundle.js"></script>
+
+  <script type="text/javascript">
+    SearchWidget.new({
+      endpoint: ""
+    })
+  </script>
+</html>
