+++ src/store/configureStore.dev.js
@@ -1,0 +1,27 @@
+import { createStore, compose, applyMiddleware } from 'redux';
+import { persistState } from 'redux-devtools';
+import rootReducer from '../reducers';
+import thunk from 'redux-thunk';
+
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+
+const enhancer = composeEnhancers(
+  applyMiddleware(thunk),
+  persistState(
+    window.location.href.match(
+      /[?&]debug_session=([^&#]+)\b/
+    )
+  )
+);
+
+export default function configureStore(initialState) {
+  const store = createStore(rootReducer, initialState, enhancer);
+
+  if (module.hot) {
+    module.hot.accept('../reducers', () =>
+      store.replaceReducer(require('../reducers').default)
+    );
+  }
+
+  return store;
+}
